refactor(assets): svg 컴포넌트화
