+++ src/global/types/SVGAttribute.type.ts
@@ -2,4 +2,6 @@ export default interface SVGAttribute {
   width?: number;
   height?: number;
   pointable?: boolean;
+  color?: string;
+  direction?: "top" | "right" | "bottom" | "left";
 }
